feat: Add a worker class
move worker class

Issue #25
